+++ packages/core/src/2d/sprite/enums/SpriteMaskUpdateFlag.ts
@@ -1,0 +1,9 @@
+/**
+ * @internal
+ */
+export enum SpriteMaskUpdateFlag {
+  /** UV. */
+  UV = 0x2,
+  /** All. */
+  All = 0x3
+}
