+++ packages/core/src/2d/sprite/enums/SpriteRendererUpdateFlag.ts
@@ -1,0 +1,9 @@
+/**
+ * @internal
+ */
+enum SpriteRendererUpdateFlag {
+  /** UV. */
+  UV = 0x2,
+  /** All. */
+  All = 0x3
+}
