+++ packages/core/src/Renderer.ts
@@ -5,6 +5,7 @@ import { assignmentClone, deepClone, ignoreClone, shallowClone } from "./clone/C
 import { Component } from "./Component";
 import { dependentComponents } from "./ComponentsDependencies";
 import { Entity } from "./Entity";
+import { RendererUpdateFlag } from "./enums/RendererUpdateFlag";
 import { Material } from "./material/Material";
 import { RenderContext } from "./RenderPipeline/RenderContext";
 import { Shader } from "./shader";
@@ -374,14 +375,4 @@ export class Renderer extends Component {
       materials[index] = material;
     }
   }
-}
-
-/**
- * @internal
- */
-export enum RendererUpdateFlag {
-  /** None. */
-  None = 0,
-  /** Include world position and world bounds. */
-  WorldVolume = 0x1
 }
