+++ packages/core/src/enums/RendererUpdateFlag.ts
@@ -1,0 +1,9 @@
+/**
+ * @internal
+ */
+export enum RendererUpdateFlag {
+  /** None. */
+  None = 0,
+  /** Include world position and world bounds. */
+  WorldVolume = 0x1
+}
