+++ packages/core/src/mesh/enums/MeshRendererUpdateFlag.ts
@@ -1,0 +1,4 @@
+export enum MeshRendererUpdateFlag {
+    VertexElements = 0x2,
+    All = 0x3
+  }
